tiny edits to setup.py
change version number to 1.1.0a (the idea is to keep this consistent
with the version of number of the public release of OQMD itself).
add Python version (2.6--2.7) compatibility info.
edit install_requires to reflect changes in package names and certain
packages which only play nice if the correct version is used.