Output.stress -> sxx, ... Edit call to pygraphviz
Details:
--------

- Calculation view threw up an error: "IndexError: only integers, slices
(`:`), ellipsis (`...`), numpy.newaxis (`None`) and integer or boolean
arrays are valid indices" for `calculation.output.stresses.0`.This is probably
due to how Django templates access variables. See
https://docs.djangoproject.com/en/1.11/ref/templates/language/#variables
The order of lookups is Dictionary, Attribute/method, Numeric index. So,
my guess is that calculation.output.stresses somehow does not return the
list of stresses but the function itself, as it is a property? Not sure.
Needs further exploration. For now, instead of accessing the list of
stresses, the template accesses the individual stress components, sxx,
syy, szz, szy, syz, szx.
- Plotting network graphs require the networkx and pygraphviz packages
to be installed. The latter in turn requires Graphviz, and the relevant
development libaries to be installed (on Debian, the packages are
graphviz-dev, libgraphviz-dev). The network graph visualizer on the
Phase Diagrams page (for # components > 4) threw an error: `networkx`
module does not have an attribute `pygraphviz_layout()`. The
documentation for the current version of networkx documents the function
call to be `networkx.nx_agraph.graphviz_layout(G)`. Not sure if this
changed with a recent version update of networkx.