Add related_name to `Site`, delete unhashable keys
Fix issues with saving Structure instances, and Dictionary keys being
unhashable objects
- In Structure.symmetrize() function, origins is now a dictionary with
site index as the key instead of the Site object itself. Because the
related Site is not always stored (it need not always be; for instance,
this function should be callable on any user Structure instances without
saving to this database), the Site instance is not hashable.
["TypeError: Model instances without primary key value are unhashable"]
- In the Structure.save() function, the site_set or Site instances need
to be assigned before atom_set or Atom instances because the Atom class
needs Site as a ForeignKey (needs to be saved).
- Related: in the declaration of model relationship of the Site class
with a Structure class, "related_name" was not set. This is fixed.