Add new PyCifRW output as test case.
NOTE: The same structure from OQMD now produces a slightly different CIF
from PyCifRW with slightly different order of symmetry operations than
output from previous versions. (I did not check thoroughly if only the
order of symmetry operations are different. If not, this can be a
serious bug in PyCifRW.)